# Add quick test whether method is event/property part