I've found a way to improve a few more places, making the call dispatch performance even better. According to the ECMA special methods have special names, so we could use that to omit expensive look ups for each call.

Benchmarking:
```
ENVIRONMENT:

BenchmarkDotNet=v0.10.13, OS=Windows 10 Redstone 2 [1703, Creators Update] (10.0.15063.786)
Intel Core i7-4770 CPU 3.40GHz (Haswell), 1 CPU, 8 logical cores and 4 physical cores
Frequency=3320308 Hz, Resolution=301.1769 ns, Timer=TSC
  [Host]     : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0
  DefaultJob : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0

BEFORE:
                            Method |     Mean |     Error |    StdDev |
---------------------------------- |---------:|----------:|----------:|
 InvokeRegularMethodInterfaceProxy | 3.853 us | 0.0764 us | 0.1096 us |
     InvokeRegularMethodClassProxy | 6.190 us | 0.1182 us | 0.1362 us |

AFTER:
                            Method |     Mean |     Error |    StdDev |
---------------------------------- |---------:|----------:|----------:|
 InvokeRegularMethodInterfaceProxy | 3.015 us | 0.0495 us | 0.0439 us |
     InvokeRegularMethodClassProxy | 5.430 us | 0.0909 us | 0.0759 us |
```

As you can see, the performance boost is ~20% without making the significant changes to the code 😄 

@alexandrnikitin @dtchepak Please take a look.