@dtchepak 2018-04-21
@zvirja Just confirming: CLI affects all .NET assemblies, correct? I noticed the [CLS] tag on points in II.22.28 -- these assumptions aren't only for CLS-compliant code?