@zvirja 2018-04-22
@dtchepak Well, yes, those naming rules describe the CLS compliant code only. However, I'm 99% sure that those compiler rules are being followed for all the major languages (c#, f#, vb.net) even if compliance flag is set to `false` . Probably, there might be difference for some exotic languages running on top of CLR, but it's unlikely - I'm pretty sure they will try to make their code as CLS compliant as possible.

I'd suggest to proceed with this optimization and see how it goes. If it happens that it breaks something (I'd be really surprised though), we can always revert the fix :blush: