@dtchepak 2018-04-23
@zvirja Thanks for the info. I've already merged it I just wanted to see if it was possible in theory for the assumptions to break down for a valid .NET assembly. 😃